Rework streaming tools into own module
