+++ seaworthy/__init__.py
@@ -8,7 +8,8 @@ seaworthy
 """
 
 from .helpers import DockerHelper
-from .logs import output_lines, wait_for_logs_matching
+from .stream.logs import wait_for_logs_matching
+from .utils import output_lines
 
 __all__ = ['DockerHelper', 'output_lines', 'wait_for_logs_matching']
