+++ seaworthy/stream/__init__.py
@@ -1,0 +1,3 @@
+from ._timeout import stream_timeout
+
+__all__ = ['stream_timeout']
