+++ seaworthy/tests-core/stream/test_logs.py
@@ -7,9 +7,8 @@ from docker.models.containers import ExecResult
 
 from seaworthy.checks import docker_client, dockertest
 from seaworthy.helpers import DockerHelper, fetch_images
-from seaworthy.logs import (
-    EqualsMatcher, OrderedLinesMatcher, RegexMatcher, UnorderedLinesMatcher,
-    output_lines, stream_logs, wait_for_logs_matching)
+from seaworthy.stream.logs import stream_logs, wait_for_logs_matching
+from seaworthy.stream.matchers import EqualsMatcher
 
 # We use this image to test with because it is a small (~4MB) image from
 # https://github.com/docker-library/official-images that we can run shell
@@ -21,217 +20,6 @@ IMG = 'alpine:latest'
 def setUpModule():  # noqa: N802 (The camelCase is mandated by unittest.)
     with docker_client() as client:
         fetch_images(client, [IMG])
-
-
-class TestOutputLinesFunc(unittest.TestCase):
-    def test_bytes(self):
-        """String lines are parsed from output bytes."""
-        self.assertEqual(output_lines(b'foo\nbar\n'), ['foo', 'bar'])
-
-    def test_exec_result(self):
-        """String lines are parsed from an ExecResult."""
-        self.assertEqual(output_lines(ExecResult(128, b'foo\r\nbar\r\n')),
-                         ['foo', 'bar'])
-
-    def test_custom_encoding(self):
-        """String lines can be parsed using a custom encoding."""
-        self.assertEqual(output_lines(b'\xe1', encoding='latin1'), ['á'])
-
-
-class TestEqualsMatcher(unittest.TestCase):
-    def test_matching(self):
-        """ Matches exactly equal strings and nothing else. """
-        matcher = EqualsMatcher('foo')
-        self.assertTrue(matcher('foo'))
-        self.assertFalse(matcher('foobar'))
-
-    def test_str(self):
-        """ The string representation is readable. """
-        matcher = EqualsMatcher('bar')
-        self.assertEqual(str(matcher), "EqualsMatcher('bar')")
-        self.assertEqual(repr(matcher), str(matcher))
-
-
-class TestRegexMatcher(unittest.TestCase):
-    def test_matching(self):
-        """
-        Matches strings that match the pattern, doesn't match other strings.
-        """
-        matcher = RegexMatcher(r'^foo')
-        self.assertTrue(matcher('foobar'))
-        self.assertFalse(matcher('barfoo'))
-
-    def test_str(self):
-        """ The string representation is readable. """
-        matcher = RegexMatcher(r'^bar')
-        self.assertEqual(str(matcher), "RegexMatcher('^bar')")
-        self.assertEqual(repr(matcher), str(matcher))
-
-
-class TestOrderedLinesMatcher(unittest.TestCase):
-    def test_matching(self):
-        """
-        Applies each matcher sequentially and returns True on the final match.
-        """
-        matcher = OrderedLinesMatcher(
-            EqualsMatcher('foo'), RegexMatcher('^bar'))
-
-        self.assertFalse(matcher('barfoo'))
-        self.assertFalse(matcher('foo'))
-        self.assertFalse(matcher('baz'))
-        self.assertTrue(matcher('barfoo'))
-
-    def test_by_equality(self):
-        """
-        The ``by_equality`` utility method takes a list of strings and produces
-        a matcher that matches those strings by equality sequentially.
-        """
-        matcher = OrderedLinesMatcher.by_equality('foo', 'bar')
-
-        self.assertFalse(matcher('bar'))
-        self.assertFalse(matcher('foo'))
-        self.assertFalse(matcher('baz'))
-        self.assertTrue(matcher('bar'))
-
-    def test_by_regex(self):
-        """
-        The ``by_regex`` utility method takes a list of patterns and produces a
-        matcher that matches by those regex patterns sequentially.
-        """
-        matcher = OrderedLinesMatcher.by_regex(r'^foo', r'bar$')
-
-        self.assertFalse(matcher('fuzzbar'))
-        self.assertFalse(matcher('foobar'))
-        self.assertFalse(matcher('baz'))
-        self.assertTrue(matcher('foobar'))
-
-    def test_exhaustion(self):
-        """
-        Once all matchers have been matched, further calls to ``match`` should
-        raise an error.
-        """
-        matcher = OrderedLinesMatcher.by_equality('foo')
-        self.assertTrue(matcher('foo'))
-
-        with self.assertRaises(RuntimeError) as cm:
-            matcher('bar')
-        self.assertEqual(str(cm.exception),
-                         'Matcher exhausted, no more matchers to use')
-
-    def test_str(self):
-        """
-        The string representation is readable and shows which matchers have
-        been matched and which are still to be matched.
-        """
-        matcher = OrderedLinesMatcher(
-            EqualsMatcher('foo'), RegexMatcher('^bar'))
-
-        self.assertEqual(
-            str(matcher),
-            'OrderedLinesMatcher(matched=[], '
-            "unmatched=[EqualsMatcher('foo'), RegexMatcher('^bar')])")
-        self.assertEqual(repr(matcher), str(matcher))
-
-        self.assertFalse(matcher('foo'))
-
-        self.assertEqual(
-            str(matcher),
-            "OrderedLinesMatcher(matched=[EqualsMatcher('foo')], "
-            "unmatched=[RegexMatcher('^bar')])")
-        self.assertEqual(repr(matcher), str(matcher))
-
-        self.assertTrue(matcher('barfoo'))
-
-        self.assertEqual(
-            str(matcher),
-            "OrderedLinesMatcher(matched=[EqualsMatcher('foo'), "
-            "RegexMatcher('^bar')], unmatched=[])")
-        self.assertEqual(repr(matcher), str(matcher))
-
-
-class TestUnorderedLinesMatcher(unittest.TestCase):
-    def test_matching(self):
-        """
-        Applies each matcher sequentially and returns True on the final match.
-        """
-        matcher = UnorderedLinesMatcher(
-            EqualsMatcher('foo'), RegexMatcher('^bar'))
-
-        self.assertFalse(matcher('barfoo'))
-        self.assertFalse(matcher('baz'))
-        self.assertTrue(matcher('foo'))
-
-        matcher = UnorderedLinesMatcher(
-            EqualsMatcher('foo'), RegexMatcher('^bar'))
-
-        self.assertFalse(matcher('foo'))
-        self.assertFalse(matcher('baz'))
-        self.assertTrue(matcher('barfoo'))
-
-    def test_by_equality(self):
-        """
-        The ``by_equality`` utility method takes a list of strings and produces
-        a matcher that matches those strings by equality sequentially.
-        """
-        matcher = UnorderedLinesMatcher.by_equality('foo', 'bar')
-
-        self.assertFalse(matcher('foo'))
-        self.assertFalse(matcher('baz'))
-        self.assertTrue(matcher('bar'))
-
-    def test_by_regex(self):
-        """
-        The ``by_regex`` utility method takes a list of patterns and produces a
-        matcher that matches by those regex patterns sequentially.
-        """
-        matcher = UnorderedLinesMatcher.by_regex(r'^foo', r'bar$')
-
-        self.assertFalse(matcher('foobar'))
-        self.assertFalse(matcher('baz'))
-        self.assertTrue(matcher('foobar'))
-
-    def test_exhaustion(self):
-        """
-        Once all matchers have been matched, further calls to ``match`` should
-        raise an error.
-        """
-        matcher = UnorderedLinesMatcher.by_equality('foo')
-        self.assertTrue(matcher('foo'))
-
-        with self.assertRaises(RuntimeError) as cm:
-            matcher('bar')
-        self.assertEqual(str(cm.exception),
-                         'Matcher exhausted, no more matchers to use')
-
-    def test_str(self):
-        """
-        The string representation is readable and shows which matchers have
-        been matched and which are still to be matched.
-        """
-        matcher = UnorderedLinesMatcher(
-            EqualsMatcher('foo'), RegexMatcher('^bar'))
-
-        self.assertEqual(
-            str(matcher),
-            'UnorderedLinesMatcher(matched=[], '
-            "unmatched=[EqualsMatcher('foo'), RegexMatcher('^bar')])")
-        self.assertEqual(repr(matcher), str(matcher))
-
-        self.assertFalse(matcher('foo'))
-
-        self.assertEqual(
-            str(matcher),
-            "UnorderedLinesMatcher(matched=[EqualsMatcher('foo')], "
-            "unmatched=[RegexMatcher('^bar')])")
-        self.assertEqual(repr(matcher), str(matcher))
-
-        self.assertTrue(matcher('barfoo'))
-
-        self.assertEqual(
-            str(matcher),
-            "UnorderedLinesMatcher(matched=[EqualsMatcher('foo'), "
-            "RegexMatcher('^bar')], unmatched=[])")
-        self.assertEqual(repr(matcher), str(matcher))
 
 
 class FakeLogsContainer:
